refactoring; added comments to run/coverage re running inside a docker container

+++ server/test/src/run.sh
@@ -1,21 +1,28 @@
 #!/bin/bash
 
-# already shelled inside docker container
-# collect trailing arguments to forward to tests
+# check already shelled inside docker container
+if [ ! -f /.dockerenv ]; then
+  echo "FAILED: run.sh is being executed outside of docker-container. See test.sh"
+  exit 1
+fi
 
-if [[ $1 == *.rb ]]; then
+# collect trailing arguments to forward to tests
+if [[ $1 == *_test.rb ]]; then
+  # shebang -> test_wrapper.sh -> test.sh -> run.sh
   FILES=($1)
   ARGS=()
 else
+  # test.sh -> run.sh
   FILES=(*_test.rb)
   ARGS=${*}
 fi
 
-# run the tests and collect coverage stats
+# run the tests with coverage
 COV_DIR=/tmp/coverage
 mkdir ${COV_DIR}
 TEST_LOG=${COV_DIR}/test.log
-# (**) These requires turn off the effect of the shebangs in each individual file
+# turn off the effect of the shebangs in each individual file
 ruby -e "%w( ${FILES[*]} ).map{ |file| './'+file }.each { |file| require file }" -- ${ARGS[@]} | tee ${TEST_LOG}
 MY_DIR="$( cd "$( dirname "${0}" )" && pwd )"
+# collect coverage stats
 ruby ${MY_DIR}/../check_test_results.rb ${TEST_LOG} ${COV_DIR}/index.html
